requirements.txt
fastapi==0.128.5
jinja2==3.1.2
python-dotenv==1.0.1
static_ffmpeg==2.13
yt-dlp==2026.2.4
uvicorn==0.40.0
